+++ sync-for-expenses/src/funcs/accountsCreate.ts
@@ -105,7 +105,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc(
     "/companies/{companyId}/connections/{connectionId}/push/accounts",
   )(pathParams);