+++ sync-for-expenses/src/funcs/accountsGetCreateModel.ts
@@ -108,7 +108,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc(
     "/companies/{companyId}/connections/{connectionId}/options/chartOfAccounts",
   )(pathParams);