+++ sync-for-expenses/src/funcs/bankAccountsGetCreateModel.ts
@@ -106,7 +106,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc(
     "/companies/{companyId}/connections/{connectionId}/options/bankAccounts",
   )(pathParams);