+++ sync-for-expenses/src/funcs/expensesCreate.ts
@@ -111,7 +111,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc(
     "/companies/{companyId}/sync/expenses/expense-transactions",
   )(pathParams);