+++ sync-for-expenses/src/funcs/manageDataGetPullOperation.ts
@@ -99,7 +99,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc("/companies/{companyId}/data/history/{datasetId}")(
     pathParams,
   );