+++ sync-for-expenses/src/funcs/pushOperationsGet.ts
@@ -100,7 +100,6 @@ async function $do(
       { explode: false, charEncoding: "percent" },
     ),
   };
-
   const path = pathToFunc("/companies/{companyId}/push/{pushOperationKey}")(
     pathParams,
   );