+++ sync-for-expenses/src/funcs/reimbursementsCreate.ts
@@ -109,7 +109,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc(
     "/companies/{companyId}/sync/expenses/reimbursable-expense-transactions",
   )(pathParams);