+++ sync-for-expenses/src/funcs/suppliersUpdate.ts
@@ -117,7 +117,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc(
     "/companies/{companyId}/connections/{connectionId}/push/suppliers/{supplierId}",
   )(pathParams);