+++ sync-for-expenses/src/funcs/transfersCreate.ts
@@ -118,7 +118,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc(
     "/companies/{companyId}/sync/expenses/transfer-transactions/{transactionId}",
   )(pathParams);